requirements.txt
asgiref==3.5.1
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.8
django-ckeditor==6.4.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-htmlmin==0.11.0
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-phonenumber-field==6.1.0
django-session-timeout==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.48
Pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
